src/main/java/teamcity/jmx/AgentBuildFilter.java
/*
 * Copyright 2018 Rod MacKenzie.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package teamcity.jmx;

import jetbrains.buildServer.serverSide.SBuild;

public class AgentBuildFilter implements BuildFilter {

    private int agentId;

    public AgentBuildFilter(int agentId) {
        this.agentId = agentId;
    }

    @Override
    public boolean accept(SBuild build) {
        return agentId == build.getAgent().getId();
    }
}
